12-27-17 B Codebase - combined SONY and CANON bases, and added 50 sec redundancy counter
Tweaked a bit of the REC/STBY timeout function, added a little
redundancy check (on sensor events, there is ALWAYS a "detected" sensor
event 50 seconds after the initial one.  I'm not sure what the reason is
for the false positive, but I added a -1 counter after 45 seconds that
ignores that extra sensor event, and it cuts down on "extra"
non-intentional events completely... or so it seems at least) - I need
to figure out *what* is causing that false positive in the end, but at
the moment, it's a bodge tradeoff that does what it's supposed to do!